feat: agrega validaciones al caso de uso

+++ questions/src/main/java/com/saber11/questions/domain/usecase/QuestionUseCase.java
@@ -12,18 +12,51 @@ import java.util.List;
 @AllArgsConstructor
 public class QuestionUseCase {
 
-    QuestionGateway questionGateway;
+    private final QuestionGateway questionGateway;
 
 
     public Question createQuestion(Question question){
+        if(question == null){
+            throw new IllegalArgumentException("Question cannot be null");
+        }
+        if (question.getArea() == null) {
+            throw new IllegalArgumentException("Area id cannot be null");
+        }
+        if (question.getQuestion() == null || question.getQuestion().trim().isEmpty()) {
+            throw new IllegalArgumentException("Question question cannot be null");
+        }
+        if (question.getAnswer()==null || question.getAnswer().trim().isEmpty()) {
+            throw new IllegalArgumentException("Answer cannot be null");
+        }
+        if (question.getOptionA()==null  || question.getOptionA().trim().isEmpty()) {
+            throw new IllegalArgumentException("Option A cannot be null");
+        }
+        if (question.getOptionB()==null || question.getOptionB().trim().isEmpty()) {
+            throw new IllegalArgumentException("Option B cannot be null");
+        }
+        if (question.getOptionC()==null  || question.getOptionC().trim().isEmpty()) {
+            throw new IllegalArgumentException("Option C cannot be null");
+        }
+        if (question.getOptionD()==null || question.getOptionD().trim().isEmpty()) {
+            throw new IllegalArgumentException("Option D cannot be null");
+        }
+        if (question.getJustification()==null || question.getJustification().trim().isEmpty()) {
+            throw new IllegalArgumentException("Justification cannot be null");
+        }
         return questionGateway.createQuestion(question);
     }
 
     public Question findQuestionById(Long id){
+        if(id == null){
+            throw new IllegalArgumentException("Question id cannot be null");
+        }
         return questionGateway.findQuestionById(id);
     }
 
     public List<Question> findQuestionsByArea(Area area){
+        if(area == null){
+            throw new IllegalArgumentException("Area cannot be null");
+        }
         return questionGateway.findQuestionsByArea(area);
     }
 
@@ -32,9 +65,42 @@ public class QuestionUseCase {
     }
 
     public Question updateQuestion(Question question){
+        if(question == null){
+            throw new IllegalArgumentException("Question cannot be null");
+        }
+        if (question.getId() == null) {
+            throw new IllegalArgumentException("Question id cannot be null");
+        }
+        if (question.getArea() == null) {
+            throw new IllegalArgumentException("Area id cannot be null");
+        }
+        if (question.getQuestion() == null || question.getQuestion().trim().isEmpty()) {
+            throw new IllegalArgumentException("Question question cannot be null");
+        }
+        if (question.getAnswer()==null || question.getAnswer().trim().isEmpty()) {
+            throw new IllegalArgumentException("Answer cannot be null");
+        }
+        if (question.getOptionA()==null  || question.getOptionA().trim().isEmpty()) {
+            throw new IllegalArgumentException("Option A cannot be null");
+        }
+        if (question.getOptionB()==null || question.getOptionB().trim().isEmpty()) {
+            throw new IllegalArgumentException("Option B cannot be null");
+        }
+        if (question.getOptionC()==null  || question.getOptionC().trim().isEmpty()) {
+            throw new IllegalArgumentException("Option C cannot be null");
+        }
+        if (question.getOptionD()==null || question.getOptionD().trim().isEmpty()) {
+            throw new IllegalArgumentException("Option D cannot be null");
+        }
+        if (question.getJustification()==null || question.getJustification().trim().isEmpty()) {
+            throw new IllegalArgumentException("Justification cannot be null");
+        }
         return questionGateway.updateQuestion(question);
     }
     public void deleteQuestionById(Long id){
+        if(id == null){
+            throw new IllegalArgumentException("Question id cannot be null");
+        }
         questionGateway.deleteQuestionById(id);
     }
 }
